+++ funcMatematicasBasicas.hpp
@@ -2,4 +2,4 @@ double factorial(double n);
 double potencia(double numero, unsigned short int potencia);
 double potenciaDouble(double numero, double potencia);
 double binomio(double a, double b);
-unsigned long long int convierteDoubleEnInt (double doble);
+unsigned int convierteDoubleEnInt (double doble);
